Mute a compiler warning about parseHandle not being initialized.
Compilers whine about the default branch of the switch statement at line 868
leaving the variable parseHandle initialized. This is impossible since the
branch of the parent switch is only taken when optIndex actually has one of the
values checked by the child switch. Let's initialize parseHandle nonetheless,
just for the sake of compilers' peace of mind.